runtime, syscall: Workaround for bug in linux's execve
linux's excve has (or had) a bug when it ran concurrently with clone,
in that it would fail to set up some datastructures if the thread
count before and after some steps differed. This is described better
and in more detail by Colin King in launchpad¹ and kernel² bugs. I
created an issue in golang's github³ for this as it first manifested
in some relatively hard to track down behaviour when exec'ing a setuid
binary from inside go, and wanted to avoid other people wasting time
tracking this down.

After some discussion there I offered to work on a workaround that
would stop go from triggering this issue, which is this.

1. https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1672819
2. https://bugzilla.kernel.org/show_bug.cgi?id=195453
3. https://github.com/golang/go/issues/19546

Change-Id: I126e87d1d9ce3be5ea4ec9c7ffe13f92e087903d