cmd/asm: Fix a bug in assembly encoding test
It is expected to test assembly code for ARMv5, ARMv6 and ARMv7
in cmd/asm/internal/asm/endtoend_test.go. But actually the loop
in "func TestARMEndToEnd(t *testing.T)" runs three times all
for ARMv5.

A new armv6.s is also added and it is tested only with GOARM=6.

Change-Id: I5dbf00809a47ace2c195335e2c9bdd768479aada