Update patch set 3
Patch Set 3:

I don't really understand what's going on with the reflection subtleties here, and don't have the time to investigate it. Even if I did, if it were just up to me, I'd say "no" because the performance gains are not sufficient for the complexity and subtlety this CL adds. It's scary stuff.

Maybe rsc or iant has an opinion.

Patch-set: 3
Reviewer: Gerrit User 5015 <5015@62eb7196-b449-3ce5-99f1-c037f21e1705>