cmd/compile: experiment w/ loop rotation as part of layout
Probably not for 1.9.  Making a CL for benchmarking on other
machines and not forgetting.

When laying out blocks, when the first transition into a
loop is detected, attempt to find a better first block
than the loop header itself.

Might make loop rotation phase unnecessary, might do a
better job.  Or not.

Change-Id: Ic92993b083bd022e75d9948773d960b0cb36bdea