Update patch set 2
Patch Set 2:

(1 comment)

> Maybe you should just disallow non absolute paths.

That sounds nice if I were designing the spec. :) But DWARF doesn't actually specify that they have to be absolute paths. Every one that I've seen is absolute and it would be silly to put a relative path in, but I think it's best to follow Postel's law here and at least try to do something rather than rejecting the whole line table. You may not be able to open the file path that comes out of the line table, but that's already true.

Patch-set: 2
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>