Update patch set 2
Patch Set 2:

Add to the list of tweaks to consider: when laying out loops outside to inside, if given an example like the one above where the inner loop has two exits, 

  +------------------+            
  |    +---------+   |
  v    v         |   |
->a -> b -> c -> d   e
  |    |    |        ^
  |    |    +--------+
  v    v
  x    y

prefer to schedule-last the exit whose target has the larger post-order number (i.e., that more likely precedes the others).  In this case y < b < c < e so prefer that exit to e (from c) be rotated to the end of the loop.  This might be a tie-breaker against loop depth, and/or it might be more robust against various combinations of exits that target headers of other inner loops.

Patch-set: 2
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>