Update patch set 1
Patch Set 1:

This certainly needs some doc changes.

I'm not sure this is a good idea.  I see that @mpvl suggested it in the issue, but it seems like an unnecessary API knob.  If your parallel tests have confusing output, fix your tests, or don't run them in parallel.

Patch-set: 1