Update patch set 1
Patch Set 1:

The confusing output stems from the fact that the === RUN line may be output at a different point than the actual output of the test. This is unavoidable without changing the semantics of parallel test. This is not necessarily a result of confusing test output.

I originally suggested adding something like a === RESUME PARALLEL line when a parallel test is continued to be run. The argument against this was that this would be too verbose. In the light of your objection, that may still be a better approach, though.
> Patch Set 1:
> 
> This certainly needs some doc changes.
> 
> I'm not sure this is a good idea.  I see that @mpvl suggested it in the issue, but it seems like an unnecessary API knob.  If your parallel tests have confusing output, fix your tests, or don't run them in parallel.

Patch-set: 1
Reviewer: Gerrit User 5182 <5182@62eb7196-b449-3ce5-99f1-c037f21e1705>