Update patch set 1
Patch Set 1:

> This certainly needs some doc changes.

Agreed, but wanted to leave that out for now since I was expecting to have to make changes on the code first.

> If your parallel tests have confusing output, fix your tests

Like Marcel says, the point is that this is not always easy. For example, if the output comes from the code you're testing (e.g. debug prints).

> or don't run them in parallel.

How? Removing all t.Parallel() lines? Having to temporarily do that again and again is cumbersome. That's what this flag/feature is for.

> I originally suggested adding something like a === RESUME PARALLEL line when a parallel test is continued to be run. The argument against this was that this would be too verbose. In the light of your objection, that may still be a better approach, though.

I'm not 100% with -parallel=-1 being a better solution than === RESUME, but I still think it's the better one. For people who don't have this problem at all, the RESUME lines will be clutter and even lead to confusion.

And for my particular case, it will also be overly verbose and complicated. Instead of getting the tests to run sequentially with each test's output in contiguous lines, I'd get the outputs split up in RESUME blocks which I would have to mentally keep track of.

I realise having -1 be a special value is adding complexity, but I feel like RESUME is no better and also adds verbosity, while doing a slightly worse job at fixing the original issue.

Patch-set: 1
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>