Update patch set 1
Patch Set 1:

An argument in favor of === RESUME/ === RUN PARALLEL is that it will also be instructive in the case of parallelism > 1.
As for the grouping of comments, as long as t.Parallel() is called at the start of test function (which is often the case), this should not be an issue. But, yeah, that is a benefit of using -1 approach.

> Patch Set 1:
> 
> Also, if more people want to pitch in and there are more arguments against this, perhaps I should close the original issue milestoned for 1.9 and open a proposal one for 1.10.

Patch-set: 1
Reviewer: Gerrit User 5182 <5182@62eb7196-b449-3ce5-99f1-c037f21e1705>