Update patch set 1
Patch Set 1:

I'm finally found time to work on this again today; sorry for the delay.

What is the usual way of iterating on a change in this project? Do I rebase to master, and amend the commit? Or do I push corrections as new commits, and merges to bring it up to date with master? Something in between?

Patch-set: 1
Reviewer: Gerrit User 20020 <20020@62eb7196-b449-3ce5-99f1-c037f21e1705>