Improved Windows syscall doc capabilities.
As per Rob Pike's proposal of freezing the syscall libraries, #4
of the problem he presented regards the lack of documentation.
This change would implement a custom doc tag, `//sysdoc`,
which would function very similarly to the standard doc
comment strings. By adding the `//sysdoc` "tag" to the generated
syscall code, when developers/programmers add new syscalls,
they can add a `//sysdoc` line with comments regarding the
syscall being generated. The goal of this change is to give
golang developers a leg up in getting the syscall documentation
cleaned up and help them succeed when implementing new calls.

As the primary parsing logic relies on scanner.Scan(), I implemented
a map[string]string to store the docs, and when a function is built,
it will tack on the doc string if it one exists. This implementation is
the easiest I could think of without rewriting most of the logic.

Syscall freeze reference: https://goo.gl/0z2wQa

Signed-off-by: Mike Lloyd <kevin.michael.lloyd@gmail.com>

Change-Id: I3e21d851cf15fc5ae1699ad34ba2ed0f7f25653a