crypto/tls: mask highest bit of X25519 public key
The field of X25519 is GF(2^255-19). If an element of this field is stored
in a 32 byte array (256 bits) there will be one remaining bit.
Therefore X25519 implementations should mask the highest bit.

This change copies another bit from the public key into the highest bit to mask it
but without an extra call to the PRNG.

Change-Id: I4fa4662b96e75a50416ca2e523ff5c80bd3812df