Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> (6 comments)
> 
> Sorry for completely dropping this.  At this point this is probably for 1.10 unless you think otherwise.

No, 1.10 is completely fine with me. The upside from this is code clarity and the risk of confusing regressions is real so it's not really a good thing to land late in the cycle (I'd actually blanked on the fact that the freeze was so close when I started working on this little series).

Patch-set: 2
Reviewer: Gerrit User 5153 <5153@62eb7196-b449-3ce5-99f1-c037f21e1705>