Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> This would also need a test before it would be mergeable, so you might as well do it now.

I'm not sure how to implement this test properly. Of course I can test whether the highest bit is not zero for certain private values but this test may fail if we change the index (e.g. from 30 to 29) in the future. If you're fine with this (may we should document this within the test), I'll implement such a test.
Should such a test live in a key_agreemenet_test.go file?

Patch-set: 2
Reviewer: Gerrit User 14805 <14805@62eb7196-b449-3ce5-99f1-c037f21e1705>