Update patch set 4
Patch Set 4: Code-Review-1

The referenced bug suggests that we're not masking the MSB of an X25519 value, but is that true? [1][2]

[1] https://github.com/golang/crypto/blob/master/curve25519/curve25519.go#L790
[2] https://github.com/golang/crypto/blob/master/curve25519/mont25519_amd64.go#L66

This change implements a GREASE[3] measure that we have in BoringSSL to exercise this aspect of the spec, but it's not clear that Go should copy this.

[3] https://tools.ietf.org/html/draft-ietf-tls-grease-00

Patch-set: 4
Reviewer: Gerrit User 7285 <7285@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1