Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> That sounds better, but I don't think it gives us any more clarity over what's there already. The question I'm trying to answer (I think) is what does "no more data is available" mean? Does that mean an io.EOF is returned? Or does that that mean there was an error or a single call where no data was written? I think that's the important thing to clarify. Maybe:
> 
> > WriteTo writes data from a data source to w. It keeps writing until it encounters an error or an empty write occurs.

Well, there isn't necessarily any empty write, either.

The WriteTo method is only going to be implemented for a type that represents some collection of data.  "No more data is available" means that all the data in that type has been written.  The precise meaning of "no more data is available" depends on the type.  In some cases it means that a Read returned io.EOF.  In other cases it means that we started with some fixed amount of data, and we wrote all of it.

Patch-set: 4