Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Oh boy, more netlink fun.
> 
> As far as I know, netlink messages are required to be aligned to 4 byte boundaries.  Everything I've ever encountered has been this way, and the netlink package I wrote assumes this behavior as well.
> 
> Then again, if there is one thing I know about netlink, it's that each family tends to break the rules in its own way.  I haven't used nfnetlink at all, but wouldn't be surprised if it didn't follow the same rules as rtnetlink or genetlink.
> 
> Can you provide the source for your statement: "Per the netlink specification..."?

Unfortunately, there is no formal specification, but conventions and recommendations for using netlink. The recommendation is indeed to use 4B alignment and padding for 4B size multiple.
However, based on my interpretation of https://www.infradead.org/~tgr/libnl/doc/core.html#_message_format and on actual packets observed:
Each message is required to start on a 4B alignment. This implies that a message may or may not have a 4B multiple for the message size, depending if there is a message following it or not. Since the next message is required to start at 4B offset, the current message must have a size that's a multiple of 4B. If there is no following it, there is no real requirement for 4B size multiple.
Using nfqueue to forward SYN packets to user space, I observed both kinds of message categories. While all started on a 4B offset alignment, some had a size which was not a multiple of 4B.

Patch-set: 3
Reviewer: Gerrit User 16965 <16965@62eb7196-b449-3ce5-99f1-c037f21e1705>