cmd/link: implement trampolines for ppc64le with ext linking
When using golang on ppc64le there have been issues for a while
now when building executables that generate extremely large text
sections.  This is due to the call instruction and the limitation
on the offset field, which is smaller than most platforms.

The original solution to this problem in golang was to split the
text section when it became too large, allowing the external (GNU)
linker to insert the necessary code to handle the long call.  That
worked fine until the another size limit for the program size was hit,
where a plt_branch is created instead of a long branch.  In that case
the plt_branch code sequence expects r2 to contain the address of the
TOC, but when golang creates dynamic executables by default
(-buildmode=exe) r2 does not always contain the address of the TOC
and as a result when building programs that reach this extremely
large size, this failure can occur.

When this problem occurred on arm, the solution was to create
trampolines to handle it. On ppc64le trampolines are generated with
internal linking, but this change implements trampolines for external
linking as well.

This is not a complete resolution for #17917, because work is needed
to remove the code to handle multiple text sections but I will leave
that for the next release.

Fixes #20497
Updates #17917

Change-Id: If5400b0f86c2c08e106b332be6db0b259b07d93d