Update patch set 2
Patch Set 2:

> Is there anything I can do to move this forward?

Mention the issue number in the comment on execlock.  Add a TODO to the comment to use some sort of rwmutex so that we don't have to serialize thread creation.  Update the overall CL comment--it still says that this should not be merged.  Mention the issue number in the overall CL comment.

Patch-set: 2