Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Making panicwrap completely argument-free allows us to avoid generating a prolog for it, which has a significant effect on the sort benchmarks. See CL 36809. Is there a way to make this fix that doesn't involve adding an argument? It's ok if panicwrap is very inefficient instead.

Ah, interesting. I didn't realize this was a fairly new thing (or that there was in fact a test for panicwrap already in the tree; is there a reason for that to be in test/ rather than a runtime test?)

Since this only requires one word (rather than six like panicwrap used to), I suspect it has much less impact. For example, the sort.(*intSort) method wrappers are still small enough with this change to not have a prologue.

Of course, I'm happy to do this a different way. I only changed panicwrap in service of the next CL. As Ian pointed out on CL 36809, panicwrap seems like a fair amount of mechanism for little gain, so we could just remove it. We could make getcallerpc an intrinsic so it doesn't need the silly dummy argument. We could expose some more internal control over the stack walk so it can see the wrapper.

Patch-set: 1
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>