Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (2 comments)
> 
> I feel my rewrite in the previous round was better. If you want to soften expectations, go with:
> 
> Work is underway on an experimental package management tool, dep <link>, to learn more about how tooling can help package management. For more about dep, see <link>.

ah yes, that does eliminate redundancies. thanks!

you don't see a need, then, to call back to the versioning problem referenced in the first paragraph of the section - it's sufficient to leave that implicit? and no need to suggest that we our aim is for dep to eventually go away?

Patch-set: 2
Reviewer: Gerrit User 10156 <10156@62eb7196-b449-3ce5-99f1-c037f21e1705>