cmd/compile: fix position info for imported constants
Previously, when we imported a typed constant, we would create an
OCONV node with the ambient lineno value, which at that time refers to
the import statement. We actually want to use the constant's own
position information instead.

This CL is mostly so the following more-invasive CL that removes the
OCONV nodes can pass toolstash.

Change-Id: I8a26c1318353fb29477c0161096baaa4fa048a67