Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (3 comments)

For Rx<<0(Ry), it is OK the check Ry in the backend, but impossible for Rx. Since the x of Rx will be encoded in the return value of "func registerShift(name string, prefix rune) int64", and the following code can not know it is Fx or Rx.

And this registerShift function is specific for arm and arm64, is also contains other platform dependent checks.

Patch-set: 2
Reviewer: Gerrit User 16935 <16935@62eb7196-b449-3ce5-99f1-c037f21e1705>