Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> I think it's trying to tell me something, but I can't quite make out what :-)
> 
> …/runtime/proc.go:309:21: write barrier prohibited by caller; acquireSudog
> 	…/runtime/sema.go:115:19: called by semacquire1
> 	…/runtime/sema.go:95:13: called by semacquire
> 	…/runtime/rwmutex.go:37:13: called by (*rwmutex).rlock
> 	…/runtime/proc.go:1644:17: called by newm

Oh. Hmm. It's telling you that newm isn't allowed to interact with the garbage collector (specifically, it runs outside the scheduler's control so GC can't bring it to a safe point). But the guts of rwmutex.rlock interact with GC, so newm isn't allowed to call rwmutex.rlock.

(A closely related problem, which the compiler isn't pointing out but is on the same path, is that acquireSudog allocates, which newm also isn't allowed to do, for much the same reason.)

I'm not actually sure what to do about this.

Patch-set: 6