Update patch set 1
Patch Set 1:

This is fine, but the runtime has snippets that do this in a lot of places (including in many of the implementations of exit, though not all). Why comment this particular one? (Perhaps because it's in runtime.main, so more visible.)

Patch-set: 1
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>