Update patch set 3
Patch Set 3:

(1 comment)

At this point we should probably punt this to Go 1.10. I'm willing to review for security but it's a bit late now. It's not clear to me that reusing tURL is proper: srcset attributes don't parse exactly the same way as URLs. The tests below try to get at whether that's true.

Patch-set: 3
CC: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>