Update patch set 1
Patch Set 1:

I'll add the conditional around it and then mail this to you for real.
This doesn't solve the problem, unfortunately, it just punts it to 1.10. Either we have to enhance the prove pass to handle the multientry case or modify preemptable loops to not require it.

Patch-set: 1
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>